# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Open Integration Workbench API
  version: 0.1.0
  description: |
    REST API for Open Integration Workbench (OIW).

    Spec ref: §21.1 (REST Endpoints), §6.2 (OpenAPI 3.1 first).

    All endpoints are prefixed with `/api/v1`. Authentication is disabled in
    local mode (localhost only); team mode uses JWT via Gitea OAuth2 (spec §21.4).

    The API is the contract between the visual designer (apps/web), the CLI,
    the MCP server, and external agents. The implementation language of the
    server may change (Python FastAPI prototype → Kotlin/Spring Boot per
    ADR-PY-002 and ADR-PY-001), but this contract is versioned and stable.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Open Integration Workbench
    url: https://github.com/hehenaice/open-integration-workbench
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://oiw.localhost
    description: Team deployment (behind Traefik)
...
paths:
...
  /api/v1/projects/{projectId}/flows:
    get:
      tags: [Flows]
      summary: List flows in a project
      operationId: listFlows
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: List of flows.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlowSummary"
  /api/v1/projects/{projectId}/flows/{flowId}:
    get:
      tags: [Flows]
      summary: Get the full IR of an integration flow
      operationId: getFlow
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FlowId"
      responses:
        "200":
          description: Flow IR.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationFlow"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Flows]
      summary: Apply typed patch operations to a flow
      description: |
        Spec ref: §12.5 (Typed Patch Format), §12.1 (all mutations go through
        typed patches — the LLM never edits files directly; the UI uses the
        same patch mechanism).

        Supported operations:
          - `addNode` — add a new node with optional diagram position
          - `removeNode` — remove a node (and its edges) by ID
          - `updateNodeConfig` — update a node's config (partial merge by default)
          - `addEdge` — add an edge between two nodes
          - `removeEdge` — remove an edge by from+to
          - `moveNode` — update a node's position in diagram.json

        The server validates:
          1. Base revision matches HEAD (if `base_revision` is provided)
          2. All referenced nodes exist
          3. No cycles are introduced
          4. No duplicate node IDs

        On success, writes `flow.yaml` and `diagram.json` back to disk
        (canonical ordering — nodes sorted by ID, edges by (from, to)).

        Returns the count of applied operations and the current revision.
      operationId: patchFlow
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FlowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchRequest"
      responses:
        "200":
          description: Patch applied successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatchResponse"
        "400":
          description: Patch validation failed (e.g. duplicate node, cycle, unknown operation).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/projects/{projectId}/resources:
    get:
      tags: [Resources]
      summary: List all resource files in a project
      description: |
        Spec ref: §11.1 (resources/ directory), §6.1 (Monaco Editor),
        §10.3 (editors/ component architecture).

        Enumerates all resource files under flows/<flow>/resources/.
        Returns path, name, resource_type, language (for Monaco), and size.
      operationId: listResources
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: List of resources.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{projectId}/resources/{resourcePath}:
    get:
      tags: [Resources]
      summary: Read a resource file
      description: |
        Spec ref: §6.1 (Monaco Editor), §12.4 (resource.read).
        Path traversal is prevented.
      operationId: getResource
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: resourcePath
          in: path
          required: true
          schema:
            type: string
          description: Resource path (e.g. flows/order-to-s4/resources/scripts/normalizeOrder.groovy).
      responses:
        "200":
          description: Resource content.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceContent"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Resources]
      summary: Write (create or update) a resource file
      description: |
        Spec ref: §6.1 (Monaco Editor), §12.4 (resource.write MCP tool),
        §11.1 (resources/ directory).

        Path traversal is prevented. Only paths under flows/<flow>/resources/
        are allowed. Parent directories are created automatically.
      operationId: writeResource
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: resourcePath
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Written resource content.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceContent"
        "400":
          description: Invalid path (outside resources/, path traversal).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/projects/{projectId}/builds:
    post:
      tags: [Builds]
      summary: Build a target-profile artifact package
      description: Compiles IR to a deterministic artifact package in `dist/`. Records digest in `.oiw/compiler.lock`. Spec §8, §4.7.
      operationId: buildArtifact
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetProfile]
              properties:
                targetProfile:
                  type: string
                  description: e.g. sap-cloud-integration-2026-07
      responses:
        "200":
          description: Build result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildResult"
        "400":
          description: Build failed (e.g. undeclared target profile).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/archive/inspect:
    post:
      tags: [Archive]
      summary: Safely inspect an archive without extracting it
      description: |
        Spec §8.2 (Safe Archive Reader), §16.1 threat 1.
        Defenses: max compressed size, max uncompressed size, max entry count,
        compression ratio cap, path-traversal rejection, symlink rejection.
      operationId: inspectArchive
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [archive]
              properties:
                archive:
                  type: string
                  format: binary
      responses:
        "200":
          description: Archive manifest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveManifest"
        "400":
          description: Archive failed safety check.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: Project identifier (lowercase kebab-case).
    FlowId:
      name: flowId
      in: path
      required: true
      schema:
        type: string
      description: Flow identifier.
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
          description: Stable error code (e.g. OIW-E001).
        details:
          type: object
          additionalProperties: true
...
    FlowSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: integer
        nodeCount:
          type: integer
        testCount:
          type: integer
        labels:
          type: object
          additionalProperties:
            type: string
    IntegrationFlow:
      type: object
      description: Full flow IR. Spec §7.2.
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          type: object
          additionalProperties: true
        spec:
          type: object
          properties:
            entrypoints:
              type: array
              items:
                type: object
                additionalProperties: true
            nodes:
              type: array
              items:
                type: object
                additionalProperties: true
            edges:
              type: array
              items:
                type: object
                additionalProperties: true
            errorHandling:
              type: object
              additionalProperties: true
            extensions:
              type: object
              additionalProperties: true
        diagram:
          type: object
          additionalProperties: true
          description: Visual layout from diagram.json. Spec §7.3 rule 4.
    PatchRequest:
      type: object
      description: |
        Typed patch request. Spec §12.5.
        WP-04 Task 6: `base_revision` is REQUIRED. The server returns
        HTTP 409 Conflict with body
        `{code: "BASE_REVISION_CONFLICT", clientRevision, serverRevision}`
        when `base_revision` is missing or does not match the current
        HEAD. The agent planner captures HEAD at planning time and
        injects it into every `flow.patch` step.
      required: [base_revision, operations]
      properties:
        base_revision:
          type: string
          description: |
            REQUIRED (WP-04 Task 6). The git HEAD sha the client based
            its patch on. The server rejects stale writes with HTTP 409.
        operations:
          type: array
          minItems: 0
          items:
            type: object
            required: [op]
            properties:
              op:
                type: string
                enum: [addNode, removeNode, updateNodeConfig, addEdge, removeEdge, moveNode]
              node:
                type: object
                description: "For addNode: {id, type, config, fidelity}"
              nodeId:
                type: string
                description: "For removeNode, updateNodeConfig, moveNode"
              from:
                type: string
                description: "For addEdge, removeEdge"
              to:
                type: string
                description: "For addEdge, removeEdge"
              condition:
                type: string
                description: "Optional edge condition (for router branches)"
              config:
                type: object
                additionalProperties: true
                description: "For updateNodeConfig: partial config to merge"
              merge:
                type: boolean
                default: true
                description: "For updateNodeConfig: if false, replace the entire config"
              position:
                type: object
                properties:
                  x:
                    type: number
                  y:
                    type: number
                description: "For addNode (optional) and moveNode (required)"
    PatchResponse:
      type: object
      properties:
        applied:
          type: integer
          description: Number of operations successfully applied.
        new_revision:
          type: string
          nullable: true
          description: The current HEAD revision (unchanged until a git commit is made).
        flow_id:
          type: string
...
    ResourceSummary:
      type: object
      description: Resource file summary for the explorer. Spec §11.1, §6.1.
      properties:
        path:
          type: string
          description: Project-relative path (e.g. flows/order-to-s4/resources/scripts/normalizeOrder.groovy).
        name:
          type: string
          description: File name only.
        resource_type:
          type: string
          enum: [groovy, xslt, xsd, json-schema, wsdl, properties, unknown]
        language:
          type: string
          description: Monaco language ID (groovy, xml, json, yaml, ini, python, shell, plaintext).
        size:
          type: integer
          description: File size in bytes.
    ResourceContent:
      type: object
      description: Full resource file content. Spec §6.1, §12.4.
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: string
        resource_type:
          type: string
        size:
          type: integer
...
    BuildResult:
      type: object
      properties:
        outDir:
          type: string
        manifestPath:
          type: string
        digest:
          type: string
          description: sha256 digest of the build. Spec §4.7.
        compilerVersion:
          type: string
        targetProfile:
          type: string
        entryCount:
          type: integer
...
    ArchiveManifest:
      type: object
      properties:
        path:
          type: string
        entryCount:
          type: integer
        compressedSize:
          type: integer
        uncompressedSize:
          type: integer
        compressionRatio:
          type: number
        digest:
          type: string
        warnings:
          type: array
          items:
            type: string
        entries:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              compressedSize:
                type: integer
              uncompressedSize:
                type: integer
              isDir:
                type: boolean